+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/appservice/serviceplan/ServicePlanComboBox.java
@@ -35,14 +35,17 @@ import com.microsoft.azure.toolkit.lib.appservice.DraftServicePlan;
 import com.microsoft.azuretools.azurecommons.helpers.NotNull;
 import com.microsoft.azuretools.azurecommons.helpers.Nullable;
 import com.microsoft.azuretools.core.mvp.model.webapp.AzureWebAppMvpModel;
-import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+
+import static com.microsoft.intellij.ui.messages.AzureBundle.message;
 
 public class ServicePlanComboBox extends AzureComboBox<AppServicePlan> {
 
@@ -54,16 +57,7 @@ public class ServicePlanComboBox extends AzureComboBox<AppServicePlan> {
     private List<PricingTier> pricingTierList = new ArrayList<>(PricingTier.getAll());
     private PricingTier defaultPricingTier = PricingTier.BASIC_B2;
 
-    @Override
-    protected String getItemText(final Object item) {
-        if (Objects.isNull(item)) {
-            return EMPTY_ITEM;
-        }
-        if (item instanceof Draft) {
-            return "(New) " + ((AppServicePlan) item).name();
-        }
-        return ((AppServicePlan) item).name();
-    }
+    private Predicate<AppServicePlan> servicePlanFilter;
 
     public void setSubscription(Subscription subscription) {
         if (Objects.equals(subscription, this.subscription)) {
@@ -98,9 +92,21 @@ public class ServicePlanComboBox extends AzureComboBox<AppServicePlan> {
         this.refreshItems();
     }
 
-    public void setValidPricingTierList(final List<PricingTier> pricingTierList, final PricingTier defaultPricingTier) {
+    public void setValidPricingTierList(@NotNull final List<PricingTier> pricingTierList, @NotNull final PricingTier defaultPricingTier) {
         this.pricingTierList = pricingTierList;
         this.defaultPricingTier = defaultPricingTier;
+        this.servicePlanFilter = appServicePlan -> pricingTierList.contains(defaultPricingTier);
+    }
+
+    @Override
+    protected String getItemText(final Object item) {
+        if (Objects.isNull(item)) {
+            return EMPTY_ITEM;
+        }
+        if (item instanceof Draft) {
+            return "(New) " + ((AppServicePlan) item).name();
+        }
+        return ((AppServicePlan) item).name();
     }
 
     @NotNull
@@ -124,6 +130,9 @@ public class ServicePlanComboBox extends AzureComboBox<AppServicePlan> {
             if (Objects.nonNull(this.os)) {
                 stream = stream.filter(p -> p.operatingSystem() == this.os);
             }
+            if (Objects.nonNull(this.servicePlanFilter)) {
+                stream = stream.filter(servicePlanFilter);
+            }
             stream = stream.sorted((first, second) -> StringUtils.compare(first.name(), second.name()));
             return stream.collect(Collectors.toList());
         }
@@ -134,7 +143,7 @@ public class ServicePlanComboBox extends AzureComboBox<AppServicePlan> {
     @Override
     protected ExtendableTextComponent.Extension getExtension() {
         return ExtendableTextComponent.Extension.create(
-            AllIcons.General.Add, "Create new app service plan", this::showServicePlanCreationPopup);
+            AllIcons.General.Add, message("appService.servicePlan.create.tooltip"), this::showServicePlanCreationPopup);
     }
 
     private void showServicePlanCreationPopup() {