+++ Utils/azuretools-core/src/com/microsoft/azuretools/sdkmanage/AzureCliAzureManager.java
@@ -34,9 +34,9 @@ import com.microsoft.azuretools.authmanage.AzureManagerFactory;
 import com.microsoft.azuretools.authmanage.CommonSettings;
 import com.microsoft.azuretools.authmanage.Environment;
 import com.microsoft.azuretools.authmanage.models.AuthMethodDetails;
-import com.microsoft.azuretools.exception.AzureRuntimeException;
 import com.microsoft.azuretools.azurecommons.helpers.Nullable;
 import com.microsoft.azuretools.enums.ErrorEnum;
+import com.microsoft.azuretools.exception.AzureRuntimeException;
 import com.microsoft.azuretools.utils.CommandUtils;
 import com.microsoft.azuretools.utils.Pair;
 import org.apache.commons.lang.ObjectUtils;
@@ -107,7 +107,7 @@ public class AzureCliAzureManager extends AzureManagerBase {
     }
 
     @Override
-    public void drop() throws IOException {
+    public void drop() {
         this.currentClientId = null;
         this.currentTenantId = null;
         this.tenantTokens.clear();
@@ -142,11 +142,7 @@ public class AzureCliAzureManager extends AzureManagerBase {
             authResult.setAzureEnv(credentials.environment().toString());
             return authResult;
         } catch (final IOException e) {
-            try {
-                drop();
-            } catch (final IOException ignore) {
-                // swallow exception while clean up
-            }
+            drop();
             throw new AzureExecutionException(FAILED_TO_AUTH_WITH_AZURE_CLI, e);
         }
     }
@@ -192,7 +188,7 @@ public class AzureCliAzureManager extends AzureManagerBase {
                                String.format(CLI_TOKEN_FORMAT_ACCESSOR_RESOURCE, tid, resource);
         final String jsonToken = CommandUtils.exec(command);
         if (StringUtils.isBlank(jsonToken)) {
-            throw new AzureRuntimeException(ErrorEnum.FAILED_TO_GET_ACCESS_TOKEN_BY_CLI);
+            throw new AzureRuntimeException(ErrorEnum.FAILED_TO_GET_ACCESS_TOKEN);
         }
         final Map<String, Object> objectMap = JsonUtils.fromJson(jsonToken, Map.class);
         final String strToken = (String) objectMap.get(CLI_TOKEN_PROP_ACCESS_TOKEN);