+++ Utils/azuretools-core/src/com/microsoft/azuretools/sdkmanage/AzureManagerBase.java
@@ -22,6 +22,7 @@
 
 package com.microsoft.azuretools.sdkmanage;
 
+import com.google.common.base.Throwables;
 import com.microsoft.azure.AzureEnvironment;
 import com.microsoft.azure.arm.resources.AzureConfigurable;
 import com.microsoft.azure.credentials.AzureTokenCredentials;
@@ -31,7 +32,9 @@ import com.microsoft.azure.management.appplatform.v2019_05_01_preview.implementa
 import com.microsoft.azure.management.resources.Subscription;
 import com.microsoft.azure.management.resources.Tenant;
 import com.microsoft.azure.toolkit.lib.common.rest.RestExceptionHandlerInterceptor;
+import com.microsoft.azuretools.adauth.AuthException;
 import com.microsoft.azuretools.authmanage.*;
+import com.microsoft.azuretools.azurecommons.helpers.Nullable;
 import com.microsoft.azuretools.enums.ErrorEnum;
 import com.microsoft.azuretools.exception.AzureRuntimeException;
 import com.microsoft.azuretools.telemetry.TelemetryInterceptor;
@@ -39,12 +42,12 @@ import com.microsoft.azuretools.utils.AzureRegisterProviderNamespaces;
 import com.microsoft.azuretools.utils.Pair;
 import org.apache.commons.lang3.StringUtils;
 
-import java.io.IOException;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Predicate;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
@@ -103,7 +106,7 @@ public abstract class AzureManagerBase implements AzureManager {
     }
 
     @Override
-    public String getTenantIdBySubscription(String subscriptionId) throws IOException {
+    public String getTenantIdBySubscription(String subscriptionId) {
         final Pair<Subscription, Tenant> subscriptionTenantPair = getSubscriptionsWithTenant().stream()
                 .filter(pair -> pair != null && pair.first() != null && pair.second() != null)
                 .filter(pair -> StringUtils.equals(pair.first().subscriptionId(), subscriptionId))
@@ -117,28 +120,41 @@ public abstract class AzureManagerBase implements AzureManager {
     }
 
     @Override
-    public List<Subscription> getSubscriptions() throws IOException {
+    public List<Subscription> getSubscriptions() {
         return getSubscriptionsWithTenant().stream().map(Pair::first).collect(Collectors.toList());
     }
 
     @Override
-    public List<Pair<Subscription, Tenant>> getSubscriptionsWithTenant() throws IOException {
+    public List<Pair<Subscription, Tenant>> getSubscriptionsWithTenant() {
         final List<Pair<Subscription, Tenant>> subscriptions = new LinkedList<>();
         final Azure.Authenticated authentication = authTenant(getCurrentTenantId());
         // could be multi tenant - return all subscriptions for the current account
         final List<Tenant> tenants = getTenants(authentication);
         for (final Tenant tenant : tenants) {
-            final Azure.Authenticated tenantAuthentication = authTenant(tenant.tenantId());
-            final List<Subscription> tenantSubscriptions = getSubscriptions(tenantAuthentication);
-            for (final Subscription subscription : tenantSubscriptions) {
-                subscriptions.add(new Pair<>(subscription, tenant));
+            try {
+                final Azure.Authenticated tenantAuthentication = authTenant(tenant.tenantId());
+                final List<Subscription> tenantSubscriptions = getSubscriptions(tenantAuthentication);
+                for (final Subscription subscription : tenantSubscriptions) {
+                    subscriptions.add(new Pair<>(subscription, tenant));
+                }
+            } catch (final Exception e) {
+                // just skip for cases user failing to get subscriptions of tenants he/she has no permission to get access token.
+                // "AADSTS50076" is the code of a weired error related to multi-tenant configuration.
+                final Predicate<Throwable> tenantError = (c) -> c instanceof AuthException && ((AuthException) c).getErrorMessage().contains("AADSTS50076");
+                if (e instanceof AzureRuntimeException && ((AzureRuntimeException) e).getCode() == ErrorEnum.FAILED_TO_GET_ACCESS_TOKEN.getErrorCode() ||
+                    Throwables.getCausalChain(e).stream().anyMatch(tenantError)) {
+                    // TODO: @wangmi better to notify user
+                    LOGGER.log(Level.WARNING, e.getMessage(), e);
+                } else {
+                    throw e;
+                }
             }
         }
         return subscriptions;
     }
 
     @Override
-    public Azure getAzure(String sid) throws IOException {
+    public @Nullable Azure getAzure(String sid) {
         if (!isSignedIn()) {
             return null;
         }
@@ -154,30 +170,26 @@ public abstract class AzureManagerBase implements AzureManager {
     }
 
     @Override
-    public AppPlatformManager getAzureSpringCloudClient(String sid) throws IOException {
+    public @Nullable AppPlatformManager getAzureSpringCloudClient(String sid) {
         if (!isSignedIn()) {
             return null;
         }
         return sidToAzureSpringCloudManagerMap.computeIfAbsent(sid, s -> {
-            String tid = this.subscriptionManager.getSubscriptionTenant(sid);
+            final String tid = this.subscriptionManager.getSubscriptionTenant(sid);
             return authSpringCloud(sid, tid);
         });
     }
 
     @Override
-    public InsightsManager getInsightsManager(String sid) throws IOException {
+    public @Nullable InsightsManager getInsightsManager(String sid) {
         if (!isSignedIn()) {
             return null;
         }
         return sidToInsightsManagerMap.computeIfAbsent(sid, s -> {
-            try {
-                // Register insights namespace first
-                final Azure azure = getAzure(sid);
-                azure.providers().register(MICROSOFT_INSIGHTS_NAMESPACE);
-            } catch (IOException e) {
-                // swallow exception while get azure client
-            }
-            String tid = this.subscriptionManager.getSubscriptionTenant(sid);
+            // Register insights namespace first
+            final Azure azure = getAzure(sid);
+            azure.providers().register(MICROSOFT_INSIGHTS_NAMESPACE);
+            final String tid = this.subscriptionManager.getSubscriptionTenant(sid);
             return authApplicationInsights(sid, tid);
         });
     }
@@ -196,7 +208,7 @@ public abstract class AzureManagerBase implements AzureManager {
     }
 
     @Override
-    public String getManagementURI() throws IOException {
+    public @Nullable String getManagementURI() {
         if (!isSignedIn()) {
             return null;
         }
@@ -217,12 +229,12 @@ public abstract class AzureManagerBase implements AzureManager {
     }
 
     @Override
-    public void drop() throws IOException {
+    public void drop() {
         LOGGER.log(Level.INFO, "ServicePrincipalAzureManager.drop()");
         this.subscriptionManager.cleanSubscriptions();
     }
 
-    protected abstract String getCurrentTenantId() throws IOException;
+    protected abstract String getCurrentTenantId();
 
     protected boolean isSignedIn() {
         return false;