+++ PluginsAndFeatures/azure-toolkit-for-intellij/resources/whatsnew/whatsnew.md
@@ -1,5 +1,32 @@
-<!-- Version: 3.44.0 -->
+<!-- Version: 3.45.0 -->
 # What's new in Azure Toolkit for IntelliJ
+
+## 3.45.0
+Flight recorder support was added in the latest Azure Toolkit for IntelliJ v3.45.0! You may monitor your online app
+ service in Azure explorer with single click and analyze the report with [Zulu Mission Control](https://www.azul.com/products/zulu-mission-control/) or IntelliJ (requires 2020.2 above Ultimate).
+
+<img src="https://user-images.githubusercontent.com/12445236/98444079-b0a79000-214a-11eb-8015-51cd2a05a1c1.gif" width="800" height="600" />
+
+Besides, file explorer is also added to Azure explorer, you could view and manage your files and logs in app service
+ with this new feature.
+
+![File Explorer](https://user-images.githubusercontent.com/12445236/98444253-93bf8c80-214b-11eb-8759-8176b520d515.png)
+
+For all the updates and fixes in this release, please refer the release notes below.
+
+### Added
+- Add file explorer for Web App and Function App in Azure explorer
+- Support flight recorder for Web App
+
+### Changed
+- New creation wizard for Function App with basic and advanced mode
+- More monitoring configuration in Web App/Function App creation wizard
+- Update template for function project
+
+### Fixed
+- [#4703](https://github.com/microsoft/azure-tools-for-java/pull/4703) Fix NPE issue in Function creation/deployment
+- [#4707](https://github.com/microsoft/azure-tools-for-java/pull/4707) Enhace error handling for azure cli token expires
+- [#4710](https://github.com/microsoft/azure-tools-for-java/pull/4710) Register service provider for insights before get insights client
 
 ## 3.44.0
 We are proud to announce the support for JBoss in IntelliJ Toolkit! You can now create JBoss Linux app service with totally new resource creation experience. You may create Web App more easily in the simple mode, while you could still set all the parameters in advance mode if necessary.